requirements.txt
prettytable~=3.11.0
colorama~=0.4.6
python-dotenv~=1.0.1
moviepy~=2.0.0.dev2
pyTelegramBotAPI~=4.22.1
pillow~=9.5.0